+++ frontend/src/components/messstelle/optionsmenue/panels/ZeitIntervall.vue
